DEV-8345 make cookie and cache interfaces as described
